Extract sync payload helpers
