+++ tests/test-custom-lens.js
@@ -307,8 +307,9 @@ assert('window exports injectLensChunks', lcSrc.includes('injectLensChunks,'));
 // ─── 12. Wiring: sync.js registration ───
 console.log('\n12. sync.js registration');
 const syncSrc = read('js/sync.js');
-assert('AI_SETTINGS_KEYS includes lens-config', syncSrc.includes("'labcharts-lens-config'"));
-assert('AI_SETTINGS_KEYS includes lens-key', syncSrc.includes("'labcharts-lens-key'"));
+const syncPayloadSrc = read('js/sync-payload.js');
+assert('AI_SETTINGS_KEYS includes lens-config', syncPayloadSrc.includes("'labcharts-lens-config'"));
+assert('AI_SETTINGS_KEYS includes lens-key', syncPayloadSrc.includes("'labcharts-lens-key'"));
 assert('ENCRYPTED_AI_KEYS includes lens-key', /ENCRYPTED_AI_KEYS[\s\S]{0,500}labcharts-lens-key/.test(syncSrc));
 
 // ─── 13. Wiring: crypto.js sensitive pattern ───
